requirements.txt
streamlit
langchain_groq 
langchain_community 
pypdf 
sentence_transformers
python-dotenv
chromadb
langchain-chroma
langchain-text-splitters
langchain-huggingface
